Refactor SuggestionList classes
